+++ .vscode/launch.json
@@ -6,6 +6,13 @@
     "configurations": [
         {
             "type": "java",
+            "name": "Launch Cafe",
+            "request": "launch",
+            "mainClass": "src.Cafe",
+            "projectName": "ECSE437-FinalProject_b490f16c"
+        },
+        {
+            "type": "java",
             "name": "Launch Current File",
             "request": "launch",
             "mainClass": "${file}"
