+++ src/Americano.java
@@ -1,11 +1,13 @@
+package src;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class Cappuccino implements Coffee {
-    private String name = "Cappuccino";
-    private double price = 4.99;
+public class Americano implements Coffee {
+    private String name = "Americano";
+    private double price = 3.99;
     private String temperature = "Hot/Cold";
-    private ArrayList<String> ingredients = new ArrayList<>(List.of("coffee bean", "water", "milk"));
+    private ArrayList<String> ingredients = new ArrayList<>(List.of("coffee bean", "water"));
 
     @Override
     public String getName() {
