+++ src/Coffee.java
@@ -1,3 +1,5 @@
+package src;
+
 import java.util.ArrayList;
 
 public interface Coffee {
