+++ src/FlatWhite.java
@@ -1,11 +1,13 @@
+package src;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class Americano implements Coffee {
-    private String name = "Americano";
-    private double price = 3.99;
+public class FlatWhite implements Coffee {
+    private String name = "Flat White";
+    private double price = 4.99;
     private String temperature = "Hot/Cold";
-    private ArrayList<String> ingredients = new ArrayList<>(List.of("coffee bean", "water"));
+    private ArrayList<String> ingredients = new ArrayList<>(List.of("coffee bean", "water", "milk"));
 
     @Override
     public String getName() {
