+++ src/Latte.java
@@ -1,8 +1,10 @@
+package src;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class FlatWhite implements Coffee {
-    private String name = "Flat White";
+public class Latte implements Coffee {
+    private String name = "Latte";
     private double price = 4.99;
     private String temperature = "Hot/Cold";
     private ArrayList<String> ingredients = new ArrayList<>(List.of("coffee bean", "water", "milk"));
