--- src/test/test.java
@@ -1,5 +1,0 @@
-package src.test;
-
-public class test {
-    
-}
